requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
requests==2.31.0
sentence-transformers==2.2.2
huggingface_hub==0.16.4
torch==2.1.0
transformers==4.35.2
numpy==1.24.4
faiss-cpu==1.7.4
openai==1.3.7
